Refactor helpers module to py_helpers
