+++ python/kwiver/vital/tests/alg/simple_bundle_adjust.py
@@ -30,7 +30,7 @@ from __future__ import print_function
 
 from kwiver.vital.algo import BundleAdjust
 from kwiver.vital.types import CameraMap
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 import numpy as np
 from multimethod import multimethod
