+++ python/kwiver/vital/tests/alg/simple_close_loops.py
@@ -30,7 +30,7 @@ from __future__ import print_function
 
 from kwiver.vital.algo import CloseLoops
 from kwiver.vital.types import CameraMap
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 
 class SimpleCloseLoops(CommonConfigurationMixin, CloseLoops):
