+++ python/kwiver/vital/tests/alg/simple_draw_detected_object_set.py
@@ -29,7 +29,7 @@
 from __future__ import print_function
 
 from kwiver.vital.algo import DrawDetectedObjectSet
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 class SimpleDrawDetectedObjectSet(CommonConfigurationMixin,
                                   DrawDetectedObjectSet):
