+++ python/kwiver/vital/tests/alg/simple_extract_descriptors.py
@@ -29,7 +29,7 @@
 from __future__ import print_function
 
 from kwiver.vital.algo import ExtractDescriptors
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 class SimpleExtractDescriptors(CommonConfigurationMixin,
                                   ExtractDescriptors):
