+++ python/kwiver/vital/tests/alg/simple_filter_features.py
@@ -29,7 +29,7 @@
 from __future__ import print_function
 
 from kwiver.vital.algo import FilterFeatures
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 class SimpleFilterFeatures(CommonConfigurationMixin,
                                   FilterFeatures):
