+++ python/kwiver/vital/tests/alg/simple_filter_tracks.py
@@ -29,7 +29,7 @@
 from __future__ import print_function
 
 from kwiver.vital.algo import FilterTracks
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 class SimpleFilterTracks(CommonConfigurationMixin,
                                   FilterTracks):
