+++ python/kwiver/vital/tests/alg/simple_transform_2d_io.py
@@ -26,7 +26,7 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 from kwiver.vital.algo import Transform2DIO
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 
 class SimpleTransform2DIO(CommonConfigurationMixin, Transform2DIO):
