+++ python/kwiver/vital/tests/alg/simple_video_input.py
@@ -31,7 +31,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 from kwiver.vital.algo import VideoInput
-from kwiver.vital.tests.helpers import CommonConfigurationMixin
+from kwiver.vital.tests.py_helpers import CommonConfigurationMixin
 
 
 class SimpleVideoInput(CommonConfigurationMixin, VideoInput):
